% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandlesID-package.R
\docType{package}
\name{RandlesID-package}
\alias{RandlesID}
\alias{RandlesID-package}
\title{RandlesID: Fast Computation of Randles Interdirections}
\description{
Fast algorithm for computing Randles interdirections
for multivariate data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ErikMendros/RandlesID}
  \item Report bugs at \url{https://github.com/ErikMendros/RandlesID/issues}
}

}
\author{
\strong{Maintainer}: Erik Mendros \email{erikmendros@gmail.com}

}
\keyword{internal}
