% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandlesID-package.R
\docType{package}
\name{RandlesID-package}
\alias{RandlesID}
\alias{RandlesID-package}
\title{RandlesID: Fast Computation of Randles Interdirections}
\description{
Fast algorithm for computing Randles interdirections
for multivariate data.
}
\author{
\strong{Maintainer}: Erik Mendros \email{erikmendros@gmail.com}

}
\keyword{internal}
